update akses login 2

+++ LogicPos/LogicPos/fkas.vb
@@ -1,35 +1,70 @@
 ﻿Imports System.Data.Odbc
 
 Public Class fkas
+    Public kodeakses As Integer
+    Dim tambahstatus, editstatus, hapusstatus As Boolean
     Dim kodekasedit As String
-    Private Sub btntambah_Click(sender As Object, e As EventArgs) Handles btntambah.Click
-        If btntambah.Text = "Tambah" Then
-            btnbatal.Enabled = True
-            btntambah.Text = "Simpan"
-            Call enable_text()
-            Call index()
-            txtkode.Text = autonumber()
-            GridControl.Enabled = False
-        Else
-            If txtkode.Text.Length = 0 Then
-                MsgBox("kode belum terisi !")
-            Else
-                If txtnama.Text.Length = 0 Then
-                    MsgBox("Nama belum terisi !")
-                Else
-                    If cbjeniskas.SelectedIndex = -1 Then
-                        MsgBox("Jenis belum terisi !")
-                    Else
-                        Call simpan()
-                    End If
-                End If
-            End If
-        End If
-    End Sub
-
     Private Sub fkas_Load(sender As Object, e As EventArgs) Handles MyBase.Load
         Me.MdiParent = fmenu
         Call awal()
+        Select Case kodeakses
+            Case 1
+                tambahstatus = True
+                editstatus = False
+                hapusstatus = False
+            Case 3
+                tambahstatus = False
+                editstatus = True
+                hapusstatus = False
+            Case 5
+                tambahstatus = False
+                editstatus = False
+                hapusstatus = True
+            Case 4
+                tambahstatus = True
+                editstatus = True
+                hapusstatus = False
+            Case 6
+                tambahstatus = True
+                editstatus = False
+                hapusstatus = True
+            Case 8
+                tambahstatus = False
+                editstatus = True
+                hapusstatus = True
+            Case 9
+                tambahstatus = True
+                editstatus = True
+                hapusstatus = True
+        End Select
+    End Sub
+    Private Sub btntambah_Click(sender As Object, e As EventArgs) Handles btntambah.Click
+        If tambahstatus.Equals(True) Then
+            If btntambah.Text = "Tambah" Then
+                btnbatal.Enabled = True
+                btntambah.Text = "Simpan"
+                Call enable_text()
+                Call index()
+                txtkode.Text = autonumber()
+                GridControl.Enabled = False
+            Else
+                If txtkode.Text.Length = 0 Then
+                    MsgBox("kode belum terisi !")
+                Else
+                    If txtnama.Text.Length = 0 Then
+                        MsgBox("Nama belum terisi !")
+                    Else
+                        If cbjeniskas.SelectedIndex = -1 Then
+                            MsgBox("Jenis belum terisi !")
+                        Else
+                            Call simpan()
+                        End If
+                    End If
+                End If
+            End If
+        Else
+            MsgBox("Tidak ada akses")
+        End If
     End Sub
 
     Sub awal()
@@ -126,14 +161,18 @@ Public Class fkas
     End Sub
 
     Private Sub btnhapus_Click(sender As Object, e As EventArgs) Handles btnhapus.Click
-        Call koneksii()
-        If MessageBox.Show("Hapus " & Me.txtnama.Text & " ?", "Konfirmasi", MessageBoxButtons.YesNo, MessageBoxIcon.Warning) = DialogResult.Yes Then
-            sql = "DELETE FROM tb_kas WHERE  kode_kas='" & txtkode.Text & "'"
-            cmmd = New OdbcCommand(sql, cnn)
-            dr = cmmd.ExecuteReader
-            MessageBox.Show(txtnama.Text + " berhasil di hapus !", "Informasi", MessageBoxButtons.OK, MessageBoxIcon.Information)
-            Me.Refresh()
-            Call awal()
+        If hapusstatus.Equals(True) Then
+            Call koneksii()
+            If MessageBox.Show("Hapus " & Me.txtnama.Text & " ?", "Konfirmasi", MessageBoxButtons.YesNo, MessageBoxIcon.Warning) = DialogResult.Yes Then
+                sql = "DELETE FROM tb_kas WHERE  kode_kas='" & txtkode.Text & "'"
+                cmmd = New OdbcCommand(sql, cnn)
+                dr = cmmd.ExecuteReader
+                MessageBox.Show(txtnama.Text + " berhasil di hapus !", "Informasi", MessageBoxButtons.OK, MessageBoxIcon.Information)
+                Me.Refresh()
+                Call awal()
+            End If
+        Else
+            MsgBox("Tidak ada akses")
         End If
     End Sub
 
@@ -215,26 +254,30 @@ Public Class fkas
     End Function
 
     Private Sub btnedit_Click(sender As Object, e As EventArgs) Handles btnedit.Click
-        If btnedit.Text = "Edit" Then
-            btnedit.Text = "Simpan"
-            btnhapus.Enabled = False
-            Call enable_text()
-            Call index()
-            GridControl.Enabled = False
-        Else
-            If txtkode.Text.Length = 0 Then
-                MsgBox("ID belum terisi !")
+        If editstatus.Equals(True) Then
+            If btnedit.Text = "Edit" Then
+                btnedit.Text = "Simpan"
+                btnhapus.Enabled = False
+                Call enable_text()
+                Call index()
+                GridControl.Enabled = False
             Else
-                If txtnama.Text.Length = 0 Then
-                    MsgBox("Nama belum terisi !")
+                If txtkode.Text.Length = 0 Then
+                    MsgBox("ID belum terisi !")
                 Else
-                    If cbjeniskas.SelectedIndex = -1 Then
-                        MsgBox("Jenis belum terisi !")
+                    If txtnama.Text.Length = 0 Then
+                        MsgBox("Nama belum terisi !")
                     Else
-                        Call edit()
+                        If cbjeniskas.SelectedIndex = -1 Then
+                            MsgBox("Jenis belum terisi !")
+                        Else
+                            Call edit()
+                        End If
                     End If
                 End If
             End If
+        Else
+            MsgBox("Tidak ada akses")
         End If
     End Sub
